refactor: Move state of selected package to service
